Feature/dapper aot

+++ src/FastGateway/Dto/StatisticLocationDto.cs
@@ -2,12 +2,12 @@
 
 public sealed class StatisticLocationDto
 {
-    public int Count { get; set; }
+    public decimal Count { get; set; }
     
     /// <summary>
     /// 归属地
     /// </summary>
     public string? Location { get; set; }
 
-    public double Ratio { get; set; }
+    public decimal Ratio { get; set; }
 }
